Split out FileAssets and then pass it in.
Changed stuff upstream so JsonLoader will let you explicitly
hand it Assets to use for loading, so this eliminates the
janky attempt to detect whether to load a file from disk or via
the standards assets approach. Which is good because, shockingly
enough, absolute paths don't start with "/" on windows, so this
didn't work.