Oh wait, no, just use FileDialog everywhere.
A little more googling shows me that it was improved in jdk7
nearly FOUR YEARS AGO. So if I run with real java, everything's
wonderful. But apparently eclipse is running things for me with
my 1.6 jre even though 1.7's my default because I've told this
project that it CAN build with that.

I'll try to ignore this since I don't normally run this from
eclipse, but don't be shocked if I bump the pom to make flumpview
require java 7 in the near future just because of this.

Right, now I'm thinking of how much fun it will be to castrate my
first bull, down on the farm.